+++ kmp-composeuiviewcontroller-ksp/src/main/kotlin/com/github/guilhe/kmp/composeuiviewcontroller/ksp/AnnotationNames.kt
@@ -4,4 +4,4 @@ private const val packageName = "com.github.guilhe.kmp.composeuiviewcontroller"
 
 internal const val composeUIViewControllerAnnotationName = "$packageName.ComposeUIViewController"
 internal const val composeUIViewControllerStateAnnotationName = "$packageName.ComposeUIViewControllerState"
-internal const val composeUIViewControllerAnnotationParameterName = "frameworkBaseName"
+internal const val frameworkBaseNameAnnotationParameter = "frameworkBaseName"