+++ sample/shared-data/src/commonMain/kotlin/com/sample/data/Components.kt
@@ -1,4 +1,4 @@
 package com.sample.data
 
-data class TestParameter(val string: String)
-data class TestParameter2(val testParameter: TestParameter)
+data class ExternalClass(val string: String)
+data class ExternalClass2(val externalClass: ExternalClass)